fix: github redirect link from github logo
