+++ src/templates/rule.js
@@ -9,6 +9,7 @@ import AccessibilityRequirements from '../components/accessibility_requirements'
 import ListOfImplementers from '../components/list-of-implementers'
 import RuleTableOfContents from '../components/rule-table-of-contents'
 
+import curateGitUrl from '../../utils/curate-git-url'
 import {
 	getGlossaryUsed,
 	getRuleUsageInRules,
@@ -221,24 +222,3 @@ export const query = graphql`
 		}
 	}
 `
-
-/**
- * Curate a given URL with
- * @param {String} url given string/ url
- * @returns {String}
- *
- * Example:
- * 		curateGitUrl("git+https://github.com/act-rules/act-rules.github.io.git")
- * yeilds
- * 		"https://github.com/act-rules/act-rules.github.io"
- */
-function curateGitUrl(url) {
-	const regexToRemove = [/^git\+/, /\.git$/]
-
-	let result = url
-	for (const regex of regexToRemove) {
-		result = result.replace(regex, '')
-	}
-
-	return result
-}
